[Issue-104] add filter for updated past month
a few notes, which I will put into the PR description as well:
This all could have been easier if you could filter based on model
properties or functions that are not fields. You cannot. Even if you try.

And I tried.

You cannot do this even if you try to put a function similar to americanhandelsociety_app.admin.Admin's
updated_past_month on americanhandelsociety_app.models.Member instead and then have your custom filter
and your Admin display field function call that property/method. What happens is that the Admin site
will 302 and redirect to <url>/<model>/?e=1, whatever that means, and not display the filter properly.

As far as I can tell, the reason for this is that the filter QUITE LITERALLY expects a filtered queryset.
NOT an exclude. Not a UNION. nothing else.

So here we are.